requirements.txt
apache-airflow == 2.4.1
apache-airflow-providers-postgres[amazon] == 5.2.2
sphinx == 5.3.0
sphinx_rtd_theme == 1.0.0